#306: Move eak card
